Add selection sort app to the virtual operational buffer approaches

+++ virtOperatBuffer/memoryLimited/ipos-gcc-vob/include/psched.h
@@ -13,10 +13,10 @@ void os_enter_critical();
 
 #define DVAR(type, var)     __nv type var
 
-#define WVAR(name, value)   wb_insert( (uint16_t) &name, value)
+#define WVAR(name, value)   wb_insert( &name, value)
 
 // If the return value of wb_get is not NULL then return it else return the value from FRAM
-#define RVAR(name)   ( (wb_get(  (uint16_t) &name)) ? __wb_get_val() : name )
+#define RVAR(name)   ( (wb_get( &name)) ? __wb_get_val() : name )
